+++ docs/src/examples/components/Loader/index.tsx
@@ -1,14 +1,18 @@
 import * as React from 'react'
 
 import Performance from './Performance'
+import Rtl from './Rtl'
 import Types from './Types'
+import Usage from './Usage'
 import Variations from './Variations'
 
 const LoaderExamples = () => (
   <>
     <Types />
     <Variations />
+    <Usage />
     <Performance />
+    <Rtl />
   </>
 )