+++ docs/src/examples/components/Menu/Types/MenuExampleUnderlined.shorthand.tsx
@@ -7,6 +7,6 @@ const items = [
   { key: 'events', content: 'Upcoming Events' },
 ]
 
-const MenuExampleUnderlined = () => <Menu defaultActiveIndex={0} items={items} underlined />
+const MenuExampleUnderlined = () => <Menu defaultActiveIndex={0} items={items} underlined primary />
 
 export default MenuExampleUnderlined