+++ packages/react/src/components/Dropdown/DropdownSearchInput.tsx
@@ -13,6 +13,11 @@ import { ComponentEventHandler, ReactProps } from '../../types'
 import { UIComponentProps } from '../../lib/commonPropInterfaces'
 import Input from '../Input/Input'
 
+export interface DropdownSearchInputSlotClassNames {
+  input: string
+  wrapper: string
+}
+
 export interface DropdownSearchInputProps extends UIComponentProps<DropdownSearchInputProps> {
   /** A dropdown search input can be formatted to appear inline in the context of a Dropdown. */
   inline?: boolean
@@ -61,9 +66,8 @@ export interface DropdownSearchInputProps extends UIComponentProps<DropdownSearc
  */
 class DropdownSearchInput extends UIComponent<ReactProps<DropdownSearchInputProps>, any> {
   static displayName = 'DropdownSearchInput'
-
   static create: Function
-
+  static slotClassNames: DropdownSearchInputSlotClassNames
   static className = 'ui-dropdown__searchinput'
 
   static propTypes = {
@@ -113,12 +117,14 @@ class DropdownSearchInput extends UIComponent<ReactProps<DropdownSearchInputProp
         onKeyUp={this.handleKeyUp}
         {...unhandledProps}
         wrapper={{
+          className: DropdownSearchInput.slotClassNames.wrapper,
           styles: styles.root,
           ...accessibilityComboboxProps,
           ...unhandledProps.wrapper,
         }}
         input={{
           type: 'text',
+          className: DropdownSearchInput.slotClassNames.input,
           styles: styles.input,
           placeholder,
           onBlur: this.handleInputBlur,
@@ -131,6 +137,11 @@ class DropdownSearchInput extends UIComponent<ReactProps<DropdownSearchInputProp
   }
 }
 
-DropdownSearchInput.create = createShorthandFactory(DropdownSearchInput)
+DropdownSearchInput.slotClassNames = {
+  input: `${DropdownSearchInput.className}__input`,
+  wrapper: `${DropdownSearchInput.className}__wrapper`,
+}
+
+DropdownSearchInput.create = createShorthandFactory({ Component: DropdownSearchInput })
 
 export default DropdownSearchInput