+++ packages/react/src/components/Popup/Popup.tsx
@@ -1,6 +1,7 @@
 import * as React from 'react'
 import * as ReactDOM from 'react-dom'
 import * as PropTypes from 'prop-types'
+import * as keyboardKey from 'keyboard-key'
 import * as _ from 'lodash'
 import { Popper, PopperChildrenProps } from 'react-popper'
 
@@ -26,7 +27,12 @@ import { getPopupPlacement, applyRtlToOffset, Alignment, Position } from './posi
 import PopupContent from './PopupContent'
 
 import { popupBehavior } from '../../lib/accessibility'
-import { FocusTrapZone, FocusTrapZoneProps } from '../../lib/accessibility/FocusZone'
+import {
+  AutoFocusZone,
+  AutoFocusZoneProps,
+  FocusTrapZone,
+  FocusTrapZoneProps,
+} from '../../lib/accessibility/FocusZone'
 
 import {
   Accessibility,
@@ -61,6 +67,9 @@ export interface PopupProps
 
   /** Initial value for 'open'. */
   defaultOpen?: boolean
+
+  /** Whether the Popup should be rendered inline with the trigger or in the body. */
+  inline?: boolean
 
   /** Delay in ms for the mouse leave event, before the popup will be closed. */
   mouseLeaveDelay?: number
@@ -141,6 +150,7 @@ export default class Popup extends AutoControlledComponent<ReactProps<PopupProps
     align: PropTypes.oneOf(ALIGNMENTS),
     defaultOpen: PropTypes.bool,
     defaultTarget: PropTypes.any,
+    inline: PropTypes.bool,
     mouseLeaveDelay: PropTypes.number,
     on: PropTypes.oneOfType([
       PropTypes.oneOf(['hover', 'click', 'focus']),
@@ -169,6 +179,7 @@ export default class Popup extends AutoControlledComponent<ReactProps<PopupProps
   private static isBrowserContext = isBrowser()
 
   private outsideClickSubscription = EventStack.noSubscription
+  private outsideKeySubscription = EventStack.noSubscription
 
   private triggerDomElement = null
   // focusable element which has triggered Popup, can be either triggerDomElement or the element inside it
@@ -179,8 +190,8 @@ export default class Popup extends AutoControlledComponent<ReactProps<PopupProps
 
   protected actionHandlers: AccessibilityActionHandlers = {
     closeAndFocusTrigger: e => {
-      this.close(e, () => _.invoke(this.triggerFocusableDomElement, 'focus'))
       e.stopPropagation()
+      this.close(e, () => _.invoke(this.triggerFocusableDomElement, 'focus'))
     },
     close: e => this.close(e),
     toggle: e => {
@@ -194,7 +205,7 @@ export default class Popup extends AutoControlledComponent<ReactProps<PopupProps
   }
 
   public componentDidMount() {
-    this.updateOutsideClickSubscription()
+    this.updateOutsideHandleSubscription()
 
     if (!this.state.open) {
       this.popupDomElement = null
@@ -202,7 +213,7 @@ export default class Popup extends AutoControlledComponent<ReactProps<PopupProps
   }
 
   public componentDidUpdate() {
-    this.updateOutsideClickSubscription()
+    this.updateOutsideHandleSubscription()
 
     if (!this.state.open) {
       this.popupDomElement = null
@@ -211,6 +222,7 @@ export default class Popup extends AutoControlledComponent<ReactProps<PopupProps
 
   public componentWillUnmount() {
     this.outsideClickSubscription.unsubscribe()
+    this.outsideKeySubscription.unsubscribe()
   }
 
   public renderComponent({
@@ -218,6 +230,7 @@ export default class Popup extends AutoControlledComponent<ReactProps<PopupProps
     rtl,
     accessibility,
   }: RenderResultConfig<PopupProps>): React.ReactNode {
+    const { inline } = this.props
     const popupContent = this.renderPopupContent(classes.popup, rtl, accessibility)
 
     return (
@@ -227,23 +240,33 @@ export default class Popup extends AutoControlledComponent<ReactProps<PopupProps
         {this.state.open &&
           Popup.isBrowserContext &&
           popupContent &&
-          ReactDOM.createPortal(popupContent, document.body)}
+          (inline ? popupContent : ReactDOM.createPortal(popupContent, document.body))}
       </>
     )
   }
 
-  private updateOutsideClickSubscription() {
+  private updateOutsideHandleSubscription() {
     if (this.state.open && this.outsideClickSubscription.isEmpty) {
       setTimeout(() => {
         this.outsideClickSubscription = EventStack.subscribe(
           'click',
           e => {
-            const isOutsidePopupElement =
-              this.popupDomElement && !this.popupDomElement.contains(e.target)
-            const isOutsideTriggerElement =
-              this.triggerDomElement && !this.triggerDomElement.contains(e.target)
+            if (this.isOutsidePopupElementAndOutsideTriggerElement(e)) {
+              this.state.open && this.trySetOpen(false, e)
+            }
+          },
+          {
+            useCapture: true,
+          },
+        )
+        this.outsideKeySubscription = EventStack.subscribe(
+          'keydown',
+          e => {
+            const keyCode = keyboardKey.getCode(e)
 
-            if (isOutsidePopupElement && isOutsideTriggerElement) {
+            const matchingKey = keyCode === keyboardKey.Enter || keyboardKey.Spacebar
+
+            if (matchingKey && this.isOutsidePopupElementAndOutsideTriggerElement(e)) {
               this.state.open && this.trySetOpen(false, e)
             }
           },
@@ -254,7 +277,15 @@ export default class Popup extends AutoControlledComponent<ReactProps<PopupProps
       })
     } else {
       this.outsideClickSubscription.unsubscribe()
+      this.outsideKeySubscription.unsubscribe()
     }
+  }
+
+  private isOutsidePopupElementAndOutsideTriggerElement(e) {
+    const isOutsidePopupElement = this.popupDomElement && !this.popupDomElement.contains(e.target)
+    const isOutsideTriggerElement =
+      this.triggerDomElement && !this.triggerDomElement.contains(e.target)
+    return isOutsidePopupElement && isOutsideTriggerElement
   }
 
   private getTriggerProps(triggerElement) {
@@ -440,7 +471,6 @@ export default class Popup extends AutoControlledComponent<ReactProps<PopupProps
       ...(rtl && { dir: 'rtl' }),
       ...accessibility.attributes.popup,
       ...accessibility.keyHandlers.popup,
-
       className: popupPositionClasses,
       style: popupPlacementStyles,
       ...this.getContentProps(),
@@ -449,13 +479,26 @@ export default class Popup extends AutoControlledComponent<ReactProps<PopupProps
     const focusTrapProps = {
       ...(typeof accessibility.focusTrap === 'boolean' ? {} : accessibility.focusTrap),
       ...popupWrapperAttributes,
+      onKeyDown: (e: React.KeyboardEvent) => {
+        // No need to propagate keydown events outside Popup
+        // when focus trap behavior is used
+        // allow only keyboard actions to execute
+        _.invoke(accessibility.keyHandlers.popup, 'onKeyDown', e)
+        e.stopPropagation()
+      },
     } as FocusTrapZoneProps
 
+    const autoFocusProps = {
+      ...(typeof accessibility.autoFocus === 'boolean' ? {} : accessibility.autoFocus),
+      ...popupWrapperAttributes,
+    } as AutoFocusZoneProps
+
     /**
-     * if there is no focus trap wrapper, we should apply
+     * if there is no focus trap  or auto focus wrapper, we should apply
      * HTML attributes and positioning to popup content directly
      */
-    const popupContentAttributes = accessibility.focusTrap ? {} : popupWrapperAttributes
+    const popupContentAttributes =
+      accessibility.focusTrap || accessibility.autoFocus ? {} : popupWrapperAttributes
 
     const popupContent = React.isValidElement(content)
       ? React.cloneElement(content, popupContentAttributes)
@@ -474,6 +517,8 @@ export default class Popup extends AutoControlledComponent<ReactProps<PopupProps
       >
         {accessibility.focusTrap ? (
           <FocusTrapZone {...focusTrapProps}>{popupContent}</FocusTrapZone>
+        ) : accessibility.autoFocus ? (
+          <AutoFocusZone {...autoFocusProps}>{popupContent}</AutoFocusZone>
         ) : (
           popupContent
         )}