+++ packages/react/src/themes/teams/components/Icon/iconStyles.ts
@@ -34,13 +34,16 @@ const getFontStyles = (
     fontFamily,
     fontSize: sizeInRems,
     lineHeight: 1,
-    textAlign: 'center',
+
+    display: 'inline-flex',
+    justifyContent: 'center',
+    alignItems: 'center',
+
+    width: sizeInRems,
+    height: sizeInRems,
 
     '::before': {
       content,
-      display: 'block',
-      width: sizeInRems,
-      height: sizeInRems,
     },
   }
 }
@@ -96,13 +99,15 @@ const iconStyles: ComponentSlotStylesInput<IconProps, IconVariables> = {
   }): ICSSInJSStyle => {
     const iconSpec = theme.icons[name]
     const rtl = theme.rtl
-    const isFontBased = !iconSpec || !iconSpec.isSvg
+    const isFontBased = name && (!iconSpec || !iconSpec.isSvg)
 
     return {
       backgroundColor: v.backgroundColor,
       display: 'inline-block',
       speak: 'none',
       verticalAlign: 'middle',
+
+      ...(!isFontBased && { boxSizing: 'border-box' }),
 
       ...(isFontBased && getFontStyles(getIconSize(size, v.sizeModifier), name)),
 
@@ -134,19 +139,19 @@ const iconStyles: ComponentSlotStylesInput<IconProps, IconVariables> = {
     }),
   }),
 
-  outlinePart: ({ variables: v }): ICSSInJSStyle => {
+  outlinePart: ({ props: p }): ICSSInJSStyle => {
     return {
       display: 'none',
 
-      ...(v.outline && {
+      ...(p.outline && {
         display: 'block',
       }),
     }
   },
 
-  filledPart: ({ variables: v }): ICSSInJSStyle => {
+  filledPart: ({ props: p }): ICSSInJSStyle => {
     return {
-      ...(v.outline && {
+      ...(p.outline && {
         display: 'none',
       }),
     }