+++ packages/react/src/themes/teams/components/Input/inputVariables.ts
@@ -33,6 +33,6 @@ export default (siteVars): InputVariables => ({
   inputPaddingWithIconAtEnd: `${pxToRem(7)} ${pxToRem(24)} ${pxToRem(7)} ${pxToRem(12)}`,
 
   inputPadding: `${pxToRem(7)} ${pxToRem(12)}`,
-  inputFocusBorderBottomColor: siteVars.brand,
+  inputFocusBorderBottomColor: siteVars.colors.primary[500],
   inputFocusBorderRadius: `${pxToRem(3)} ${pxToRem(3)} ${pxToRem(2)} ${pxToRem(2)}`,
 })