+++ packages/react/src/themes/teams/siteVariables.ts
@@ -31,7 +31,11 @@ export const red08 = '#F3D6DB' // no mapping color
 //
 // SHADOW LEVELS
 //
-export const shadowLevel1 = '0 .2rem .4rem -.075rem rgba(0, 0, 0, 0.1)'
+export const shadowLevel1 = '0 .2rem .4rem -.075rem rgba(0, 0, 0, .1)'
+export const shadowLevel2 = '0 .4rem 0.7rem -0.1rem rgba(0, 0, 0, .1)'
+export const shadowLevel3 = '0 .8rem 1rem -0.2rem rgba(0, 0, 0, .1)'
+export const shadowLevel4 = '0 .6rem 1.8rem -0.4rem rgba(0, 0, 0, .1)'
+
 export const shadowLevel1Darker = '0 .2rem .4rem -.075rem rgba(0, 0, 0, 0.5)'
 
 //